+++ tests/Modules/ModuleWithProductModel/Models/ProductTypeProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ProductTypeProxy class.
+ *
+ * @copyright   Copyright (c) 2020 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2020-12-08
+ *
+ */
+
+namespace Konekt\Concord\Tests\Modules\ModuleWithProductModel\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class ProductTypeProxy extends ModelProxy
+{
+}